[ANI-68-ani-path-detail] refactor: 타입을 정리합니다.
