+++ apps/web/app/path/[pathId]/load/page.tsx
@@ -1,12 +1,11 @@
 import { getPathLoad } from '@/views/path/api/get-path-load'
 import { notFound, redirect } from 'next/navigation'
-import { Suspense } from 'react'
 
 type Props = {
   params: { pathId: string }
 }
 
-async function PathLoader({ pathId }: { pathId: string }) {
+export default async function Page({ params: { pathId } }: Props) {
   const data = await getPathLoad({ pathId })
   const { status, data: pathData } = data
 
@@ -15,12 +14,4 @@ async function PathLoader({ pathId }: { pathId: string }) {
   }
 
   redirect(`/path/${pathId}/detail?partId=${pathData.partId}&pageId=${pathData.pageId}`)
-}
-
-export default async function Page({ params: { pathId } }: Props) {
-  return (
-    <Suspense fallback={<div>Loading...</div>}>
-      <PathLoader pathId={pathId} />
-    </Suspense>
-  )
 }
